requirements.txt
pandas
numpy
matplotlib
opencv-python
torch
torchvision
scikit-learn
pycairo
tqdm
lmdb
aiofiles